% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-nombres-normalizar.R
\name{nombre_title_case}
\alias{nombre_title_case}
\title{Poner nombres en Title Case respetando partículas}
\usage{
nombre_title_case(texto)
}
\arguments{
\item{texto}{Vector de caracteres con nombres.}
}
\value{
Vector de caracteres con formato Title Case + partículas en minúscula.
}
\description{
Convierte un nombre a "Title Case" (primera letra en mayúscula) y
luego pasa a minúscula las partículas comunes (con \code{\link[=ajustar_particulas]{ajustar_particulas()}}).
}
\keyword{internal}
