% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FW-perfil_incompleto.R
\name{perfil_incompleto}
\alias{perfil_incompleto}
\title{Perfil según flag}
\usage{
perfil_incompleto(data, flag, value = FALSE)
}
\arguments{
\item{data}{data.frame con columna \code{email} y el flag lógico correspondiente
(ej. \code{flag_long}, \code{flag_pre}, \code{flag_post}).}

\item{flag}{nombre del flag (string).}

\item{value}{valor lógico que define el perfil (por defecto \code{FALSE}).}
}
\value{
Subconjunto de \code{data} con las personas que cumplen la condición.
}
\description{
Devuelve las personas que cumplen cierta condición lógica sobre un flag
(ej. PRE sin LONG, LONG sin PRE, POST sin PRE, etc.).
}
\examples{
# PRE sin LONG
pre <- add_flags(pre, long, post)$pre
perfil_incompleto(pre, flag = "flag_long", value = FALSE)

# LONG sin PRE
long <- add_flags(pre, long, post)$long
perfil_incompleto(long, flag = "flag_pre", value = FALSE)

# POST sin PRE
post <- add_flags(pre, long, post)$post
perfil_incompleto(post, flag = "flag_pre", value = FALSE)
}
