% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizar-nombre.R
\name{ajustar_particulas}
\alias{ajustar_particulas}
\title{Ajustar partículas en nombres}
\usage{
ajustar_particulas(nombre)
}
\arguments{
\item{nombre}{Vector de caracteres con nombres.}
}
\value{
Vector de caracteres con partículas en minúscula.
}
\description{
Convierte a minúscula partículas frecuentes en nombres y apellidos
(ej. "de", "del", "la", "y", "van", "von").
}
\examples{
ajustar_particulas("Antonio De La Vega")
ajustar_particulas(c("JUAN DEL RÍO", "María VAN DER Meer"))
}
