% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-normalizar-nombre.R
\name{ajustar_particulas}
\alias{ajustar_particulas}
\title{Ajustar partículas en nombres}
\usage{
ajustar_particulas(texto)
}
\arguments{
\item{texto}{Vector de caracteres con nombres.}
}
\value{
Vector de caracteres con partículas en minúscula.
}
\description{
Ajustar partículas en nombres
}
\keyword{"de",}
\keyword{"del",}
\keyword{"la",}
\keyword{"van",}
\keyword{"von").}
\keyword{"y",}
\keyword{(ej.}
\keyword{Convierte}
\keyword{a}
\keyword{apellidos}
\keyword{en}
\keyword{frecuentes}
\keyword{internal}
\keyword{minúscula}
\keyword{nombres}
\keyword{partículas}
\keyword{y}
