% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-check-cols.R
\name{check_cols_long}
\alias{check_cols_long}
\title{Validar columnas de LONG}
\usage{
check_cols_long(long)
}
\arguments{
\item{long}{data.frame con \code{email} y \code{date}}
}
\value{
(invisible) NULL
}
\description{
Requisitos:
\itemize{
\item data.frame/tibble
\item columnas: \code{email}, \code{date}
\item \code{email} character; \code{date} clase Date o POSIXt
\item \code{email} sin vacíos/NA
}
}
