% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-check-cols.R
\name{check_cols_post}
\alias{check_cols_post}
\title{Validar columnas de POST}
\usage{
check_cols_post(post)
}
\arguments{
\item{post}{data.frame con \code{email} y \code{name}}
}
\value{
(invisible) NULL
}
\description{
Mismas reglas que PRE:
\itemize{
\item data.frame/tibble
\item columnas: \code{email}, \code{name}
\item \code{email} character y sin vacíos/NA
}
}
