% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-check-cols.R
\name{check_cols_pre}
\alias{check_cols_pre}
\title{Validar columnas de PRE}
\usage{
check_cols_pre(pre)
}
\arguments{
\item{pre}{data.frame con \code{email} y \code{name}}
}
\value{
(invisible) NULL
}
\description{
Requisitos:
\itemize{
\item data.frame/tibble
\item columnas: \code{email}, \code{name}
\item \code{email} character y sin vacíos/NA
}
}
