% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-normalizar-email.R
\name{emails_check}
\alias{emails_check}
\title{Chequea y reporta errores comunes en emails}
\usage{
emails_check(texto)
}
\arguments{
\item{texto}{Vector de emails}
}
\value{
Tibble con columnas original, normalizado, tipo_error
}
\description{
Chequea y reporta errores comunes en emails
}
