% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-google-sheet.R
\name{gs_auth}
\alias{gs_auth}
\title{Autenticación con Google Sheets (JSON de servicio)}
\usage{
gs_auth(path)
}
\arguments{
\item{path}{Ruta al JSON de credenciales.}
}
\value{
(invisible) TRUE
}
\description{
Autenticación con Google Sheets (JSON de servicio)
}
