% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ejemplo.R
\docType{data}
\name{long}
\alias{long}
\title{Datos de ejemplo: longitudinal}
\format{
Un tibble con m filas y 2 variables:
\describe{
\item{email}{Dirección de correo (con posibles errores)}
\item{date}{Fecha de la medición}
}
}
\usage{
long
}
\description{
Conjunto ficticio de mediciones longitudinales con múltiples observaciones por persona,
fechas distribuidas en varias semanas, abandonos y participantes externos sin pretest.
}
\examples{
data(long)
dplyr::count(long, email)
}
\keyword{datasets}
