% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizar-email.R
\name{normalizar_email}
\alias{normalizar_email}
\title{Normaliza emails (minúsculas, sin acentos, sin espacios)}
\usage{
normalizar_email(x)
}
\arguments{
\item{x}{Vector de emails}
}
\value{
Vector con emails normalizados
}
\description{
Normaliza emails (minúsculas, sin acentos, sin espacios)
}
