% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizar-email.R
\name{normalizar_email}
\alias{normalizar_email}
\title{Normalizar emails y detectar errores comunes}
\usage{
normalizar_email(email_vector)
}
\arguments{
\item{email_vector}{Vector de caracteres con emails.}
}
\value{
data.frame con columnas: \code{original}, \code{normalizado}, \code{tipo_error}.
}
\description{
Limpia espacios, pasa a minúsculas, elimina tildes y espacios internos.
Además etiqueta si el original tenía tildes, ñ, espacios o mayúsculas.
}
\examples{
normalizar_email(c("  JOSÉ.PeRez @GmAil.com ", "ana_ñ@example.es"))
}
