% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizar-nombre.R
\name{normalizar_nombre}
\alias{normalizar_nombre}
\title{Normalizar nombres (Primera letra en mayúscula,
excepto las partículas más comunes de algunos apellidos)}
\usage{
normalizar_nombre(nombre)
}
\arguments{
\item{nombre}{Vector de caracteres con nombres.}
}
\value{
Vector normalizado.
}
\description{
Normalizar nombres (Primera letra en mayúscula,
excepto las partículas más comunes de algunos apellidos)
}
\examples{
# Normaliza mayúsculas y espacios extra:
normalizar_nombre("  MARÍA  DEL   PILAR ")
#> [1] "María del Pilar"

# Normaliza partículas comunes:
normalizar_nombre(c("antonio de la vega", "JUAN DEL RÍO"))
#> [1] "Antonio de la Vega" "Juan del Río"
}
