% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ejemplo.R
\docType{data}
\name{post}
\alias{post}
\title{Datos de ejemplo: postest}
\format{
Un tibble con p filas y 2 variables:
\describe{
\item{email}{Dirección de correo}
\item{date}{Fecha de respuesta}
}
}
\usage{
post
}
\description{
Conjunto ficticio de postest aplicado a un subconjunto de participantes,
incluyendo casos con y sin pretest/longitudinal.
}
\examples{
data(ejemplo_post)
nrow(ejemplo_post)
}
\keyword{datasets}
