% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizar-email.R
\name{quitar_acentos}
\alias{quitar_acentos}
\title{Quitar acentos (transliterar a ASCII)}
\usage{
quitar_acentos(x)
}
\arguments{
\item{x}{Vector de caracteres.}
}
\value{
Vector sin tildes ni diacríticos.
}
\description{
Quitar acentos (transliterar a ASCII)
}
