% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-normalizar-email.R
\name{quitar_acentos}
\alias{quitar_acentos}
\title{Quita tildes/acentos y normaliza caracteres a ASCII}
\usage{
quitar_acentos(texto)
}
\arguments{
\item{texto}{Vector de caracteres.}
}
\value{
Un vector de caracteres sin acentos ni caracteres no imprimibles.
}
\description{
Convierte caracteres acentuados a su forma base y elimina
caracteres no imprimibles. Usamos \code{stringi::stri_trans_general("Latin-ASCII")}
por su mayor consistencia frente a \code{iconv} entre plataformas.
}
\examples{
quitar_acentos(c("Camión", "AÑA", "José  "))
}
